Utilities: Move ConvertIntToBoardFields from Generator to Utilities
